% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keep_clean_only.R
\name{keep_clean_only}
\alias{keep_clean_only}
\title{keep_clean_only}
\usage{
keep_clean_only(data, qual_ = 40)
}
\arguments{
\item{data}{a dataset, usually the output of \code{read_phruscle}}

\item{qual_}{the median quality to discriminate poor quality sequences.}
}
\value{
a data frame filtered off its dirty seq
}
\description{
Discard the dirty sequence
}
\author{
Samuel Barreto
}

