% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_gcbiasr_theme.R
\name{legend_position}
\alias{legend_position}
\title{legend_position}
\usage{
legend_position(x = NULL, y = NULL)
}
\arguments{
\item{x}{the x coordinate. must be between 0 and 1}

\item{y}{idem.}
}
\value{
a theme element.
}
\description{
Easily place the legend on the graph.
}
\details{
If both x and y are null, place it at the bottom by default.
}
\author{
Samuel Barreto
}

