% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_align.R
\name{plot_align}
\alias{plot_align}
\title{plot_align}
\usage{
plot_align(data, mutant_)
}
\arguments{
\item{data}{a data.frame. it must be the output of \code{read_phruscle}.}

\item{mutant_}{the mutant to filter by.}
}
\value{
a ggplot output of the alignment
}
\description{
Plot the alignment of the markers positions
}
\author{
Samuel Barreto
}

