% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_align.R
\name{plot_align_qual}
\alias{plot_align_qual}
\title{plot_align_qual}
\usage{
plot_align_qual(data, mutant_)
}
\arguments{
\item{data}{the output of \code{run_phruscle}.}

\item{mutant_}{the mutant to filter by.}
}
\value{
a ggplot2 plot.
}
\description{
Plot the alignment and the quality of the sequence juxtaposed
}
\author{
Samuel Barreto
}

