% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_qual.R
\name{plot_qual}
\alias{plot_qual}
\title{plot_qual}
\usage{
plot_qual(data, mutant_)
}
\arguments{
\item{data}{the output of \code{run_phruscle}.}

\item{mutant_}{the mutant to filter by.}
}
\value{
a ggplot2 plot.
}
\description{
Plot the quality of the sequence
}
\author{
Samuel Barreto
}

