% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_phruscle.R
\name{read_phruscle}
\alias{read_phruscle}
\title{read_phruscle}
\usage{
read_phruscle(path = NULL, mutant_levels = c("s", "w", "sw", "ws"))
}
\arguments{
\item{path}{the path to the csv file.}
}
\value{
a tbl_df
}
\description{
Read the csv dataset in ``path'' and return it as a clean dataset.
}
\author{
Samuel Barreto
}

