% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_gcbiasr_theme.R
\name{set_gcbiasr_theme}
\alias{set_gcbiasr_theme}
\title{set_gcbiasr_theme}
\usage{
set_gcbiasr_theme()
}
\value{
the default theme for ggplot2 plots
}
\description{
Set the default theme for ggplot2 plots.
}
\details{
Inspired both by the fivethirtyeight theme and the solarized theme of
ggthemes.
}
\author{
Samuel Barreto
}

